Change XTarget and ZTarget types from byte to short

+++ TombEngine/Objects/TR5/Emitter/tr5_bats_emitter.h
@@ -12,8 +12,8 @@ struct BatData
 	short Velocity;
 	short Counter;
 	short LaraTarget;
-	byte XTarget;
-	byte ZTarget;
+	short XTarget;
+	short ZTarget;
 
 	byte Flags;
